split ring_ops.py
